Trackig-main/requirements.txt
requests>=2.31.0
python-telegram-bot==20.3